requirements.txt
Django==2.2
pip
flake8
shortuuid==0.5.0
celery==4.4.6
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-bulk-update==2.2.0
django-mysql-geventpool==0.2.5
djangorestframework==3.7.7
djangorestframework-jwt==1.10.0
redis==3.5.1
celery[redis]==4.4.6
requests==2.24.0
mysqlclient==2.0.1
gunicorn==20.0.4
jenkinsapi==0.3.11
flower==0.9.2
tornado==5.1.1
pyjwt==1.7.1
jira ==2.0.0
openpyxl
mongoengine==0.21.0
xmind2testcase==1.5.0
fabric3
django-rest-framework-mongoengine==3.4.1
meican~=0.4.3
numpy==1.19.2
pandas==1.1.0
arrow==0.17.0
backoff==1.10.0
retry~=0.9.2
django_extensions==3.1.1
chinesecalendar==1.5.0
gevent==20.12.1
django-cors-headers==3.7.0
python-gitlab==2.7.1
nacos-sdk-python
clickhouse_driver
funcy==1.16
peewee==3.14.4
hutils==1.0.14
python-dateutil~=2.8.2
pymongo~=3.12.0
pytz~=2021.1
XMind~=1.2.0
PyMySQL~=1.0.2
kombu~=4.6.11

